Q5 changes

+++ src/LW_01/Circle.java
@@ -7,11 +7,11 @@ public class Circle {
         this.radius = radius;
     }
 
-    public int computeArea() {
-        return (int) (3.14 * radius * radius);
+    public double computeArea() {
+        return 3.14 * radius * radius;
     }
 
-    public int computeCircumference() {
-        return (int) (2 * 3.14 * radius);
+    public double computeCircumference() {
+        return 2 * 3.14 * radius;
     }
 }
